2.0.2: - graph: Fixed potential buffer overflow.        - audio_mixer: Added volume graph for debugging.        - write_frame: Cosmetics.        - misc: Fixed AUTO bugs caused by arethmetic overflows.        - ffmpeg_producer: Fixed nb_frames when disabling looping.
git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.0.2@1796 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ core/producer/transition/transition_producer.cpp
@@ -98,7 +98,7 @@ struct transition_producer : public frame_producer
 		return disable_audio(last_frame_);
 	}
 
-	virtual int64_t nb_frames() const override
+	virtual uint32_t nb_frames() const override
 	{
 		return get_following_producer()->nb_frames();
 	}
